requirements.txt
pytest
boto3
vcrpy
